chang run.sh to train.sh && remove show.sh

+++ train.sh
@@ -5,6 +5,14 @@ QTABLE="qtable.txt"
 VISITS="visits.txt"
 DATA="pp"
 
+DIR="$1"
+
+if [ -z $DIR ]; then
+    echo "Error: no working directory specilized"
+    echo "Usage: $0 [dir]"
+    exit
+fi
+
 rm -f $QTABLE
 rm -fr $HISTORY
 rm -f $DATA
